Refactor: LoginViewController 에서 UI 관련 코드 분리 작업
